add support for XmlAttribute to ToParentMixin
